+++ pipfreeze/__init__.py
@@ -1,1 +1,2 @@
 from .pipfreeze import PipFreeze  # flake8: noqa
+from .package import Package, Requirement  # flake8: noqa