+++ CHANGES.rst
@@ -1,3 +1,12 @@
+Unreleased
+==========
+
+* implement ``Requirement``
+* implement ``Package``
+* implement ``PipFreeze.__contains__()``
+* implement ``PipFreeze.__len__()``
+* drop py26 support
+
 0.0.2
 =====